+++ packages/@tinacms/cli/package.json
@@ -104,6 +104,7 @@
     "progress": "^2.0.3",
     "prompts": "^2.4.1",
     "readable-stream": "^4.3.0",
+    "rollup-plugin-visualizer": "^5.9.0",
     "tailwindcss": "^3.2.7",
     "url-pattern": "^1.0.3",
     "vite": "^4.1.4",
