+++ packages/@tinacms/cli/src/next/commands/dev-command/server/index.ts
@@ -4,8 +4,7 @@ import cors from 'cors'
 import { resolve as gqlResolve, Database } from '@tinacms/graphql'
 import { tinaTailwind } from '../tailwind'
 import { ConfigManager } from '../../../config-manager'
-import { createMediaRouter } from './routes'
-import { parseMediaFolder } from './models/media'
+import { parseMediaFolder, createMediaRouter } from './media'
 import { transform as esbuildTransform } from 'esbuild'
 import { pipeline } from 'readable-stream'
 import { createServer } from 'net'
